C third stage, create FAT32 image, ATA disk read
make now puts the file sizes of each of the stages into build/log.txt. The sizes and start sectors of each stage is manually set in the assembly files as constants, so these must be changed accordingly if/as the size of stage 2 or 3 increases. Sector size is assumed to be 512 bytes. Constants to be concerned about are in boot/stage2.asm.

ATA read functions from osdev work, but specifying the number of sectors to read does not do anything (it still reads just 1 sector). So i just put it in a loop to read some number of sectors more than 1 lol

Also added some basic print functions to stage3.c.

tools/fatcreate.c was modified to work with macos with changes on lines 489, 498, 504. i haven't tried, but it will likely not build on linux anymore so we gotta fix that.